+++ analysis_options.yaml
@@ -7,6 +7,9 @@
 
 # The following line activates a set of recommended lints for Flutter apps,
 # packages, and plugins designed to encourage good coding practices.
+analyzer:
+  errors:
+    must_be_immutable: ignore
 include: package:flutter_lints/flutter.yaml
 
 linter:
