+++ lib/modules/inventory/inventory_provider.dart.dart
@@ -164,40 +164,6 @@ class InventoryProvider with ChangeNotifier {
     notifyListeners();
   }
 
-  Stream<DocumentSnapshot<Map<String, dynamic>>> getTodaysOrders(
-      String businessID) {
-    String todayDateString = DateFormat('yyyy-MM-dd').format(DateTime.now());
-    try {
-      return _firestore
-          .collection('businesses')
-          .doc(businessID)
-          .collection('orders')
-          .doc(todayDateString)
-          .snapshots();
-    } catch (e) {
-      return Stream.empty();
-    }
-  }
-
-  double getOrderTotal(List<dynamic> items) {
-    print(items);
-    double total = 0;
-    for (Map<String, dynamic> item in items) {
-      if (item['status'] != 'canceled') {
-        total += item['unitPrice'] * item['quantity'];
-      }
-    }
-    return total;
-  }
-
-  Widget getOrderStatus(List<dynamic> items) {
-    String status = items.first['status'];
-    return Text(
-      status,
-      style: TextStyle(color: status == "canceled" ? Colors.red : Colors.green),
-    );
-  }
-
   void setSelectedCategory(String category) {
     selectedCategory = category;
     notifyListeners();
@@ -299,103 +265,6 @@ class InventoryProvider with ChangeNotifier {
   }
 
 //orders
-  Future<void> completeOrder(String businessID) async {
-    notifyListeners();
-    for (ProductModel p in openOrder.keys) {
-      await decreaseProductQuantity(p, openOrder[p]!, businessID);
-    }
-    await addOrderToHistory(businessID, "fulfilled");
-    openOrder.clear();
-    notifyListeners();
-  }
-
-  Map<ProductModel, int> fetchOpenOrder() {
-    return openOrder;
-  }
-
-  Future<void> addOrderToHistory(String businessID, String status) async {
-    String todayDateString = DateFormat('yyyy-MM-dd').format(DateTime.now());
-    String position = "1";
-    Map<String, List<Map<String, dynamic>>> order = {};
-    final DocumentSnapshot<Map<String, dynamic>> documentSnapshot =
-        await _firestore
-            .collection("businesses")
-            .doc(businessID)
-            .collection("orders")
-            .doc(todayDateString)
-            .get();
-
-    if (documentSnapshot.exists) {
-      position = (documentSnapshot.data()!.length + 1).toString();
-      for (ProductModel p in openOrder.keys) {
-        if (!order.containsKey(position)) {
-          order[position] = [];
-        }
-        order[position]!.add({
-          "product": p.name,
-          "quantity": openOrder[p],
-          "unitPrice": p.unitPrice,
-          "totalPrice": p.unitPrice * openOrder[p]!,
-          "status": status
-        });
-        // position = (int.parse(position) + 1).toString();
-      }
-      print(order);
-      _firestore
-          .collection("businesses")
-          .doc(businessID)
-          .collection("orders")
-          .doc(todayDateString)
-          .update(order);
-    } else {
-      for (ProductModel p in openOrder.keys) {
-        if (!order.containsKey(position)) {
-          order[position] = [];
-        }
-        order[position]!.add({
-          "product": p.name,
-          "quantity": openOrder[p],
-          "unitPrice": p.unitPrice,
-          "totalPrice": p.unitPrice * openOrder[p]!,
-          "status": status
-        });
-        // position = (int.parse(position) + 1).toString();
-      }
-      _firestore
-          .collection("businesses")
-          .doc(businessID)
-          .collection("orders")
-          .doc(todayDateString)
-          .set(order);
-    }
-  }
-
-  Future<void> clearOpenOrder(String businessID) async {
-    await addOrderToHistory(businessID, "canceled");
-    openOrder.clear();
-    notifyListeners();
-  }
-
-  setQuantityError(String? error) {
-    quantityError = error;
-    notifyListeners();
-  }
-
-  addToOrder(ProductModel p, int quantity) {
-    if (openOrder.containsKey(p)) {
-      openOrder[p] = openOrder[p]! + quantity;
-    } else {
-      openOrder[p] = quantity;
-    }
-    notifyListeners();
-  }
-
-  bool verifyQuantity(ProductModel p, int quantity) {
-    if (p.quantity >= quantity) {
-      return true;
-    }
-    return false;
-  }
 
   int calculateAboveCriticalLevel() {
     int aboveCriticalLevel = 0;
@@ -540,7 +409,7 @@ class InventoryProvider with ChangeNotifier {
   }
 
   Future<void> fetchProducts(String businessID) async {
-    // cleanDB(businessID);
+    print(isProductFetched);
     if (!isProductFetched) {
       allProdcuts = [];
       try {
